refactor: 将数值类型从u16/i16改为usize/isize
统一游戏中的数值类型为usize/isize以提高一致性和减少类型转换
添加符号缓存以提高渲染性能
改进AI决策逻辑，考虑可达空间

+++ src/app.rs
@@ -43,8 +43,8 @@ fn calculate_game_size(width: u16, height: u16) -> Option<GameSize> {
         return None;
     }
     Some(GameSize {
-        width: width - STATUS_WIDTH - 2,
-        height: height - TITLE_HEIGHT - FOOTER_HEIGHT - 2,
+        width: (width - STATUS_WIDTH - 2) as usize,
+        height: (height - TITLE_HEIGHT - FOOTER_HEIGHT - 2) as usize,
     })
 }
